% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxplot_features}
\alias{boxplot_features}
\title{Generates a boxplot with the distribution of the values across the GTEx tissues, including description of features}
\usage{
boxplot_features(values, metadata, exons, junctions)
}
\arguments{
\item{values}{data.frame with the values. rownames need to be the feature identifier (gene_id, exon_id,...) and colnames the sample ids.}

\item{metadata}{object returned by load_metadata()}

\item{exons}{exons}

\item{junctions}{junctions}
}
\description{
Generates a boxplot with the distribution of the values across the GTEx tissues, including description of features
}
\seealso{
\code{\link{load_metadata}}
}

