\name{ggtex-package}
\alias{ggtex-package}
\alias{ggtex}
\docType{package}
\title{
ggtex
}
\description{
Set of tools for visualization of GTEx data
}
\details{
\tabular{ll}{
Package: \tab ggtex\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2015-06-10\cr
License: \tab GPL-2|GPL-3\cr
}
The ggtex package provides with the following functions:

load_metadata() loads the GTEx metadata given the correspondig path to the files
ggtex_boxplot_tissues() give a matrix of values, plot the distribution across tissues
}
\author{
Didac Santesmasses

Maintainer: <didac.santesmasses@crg.cat>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
