% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggtex_boxplot_tissues}
\alias{ggtex_boxplot_tissues}
\title{Generates a boxplot with the distribution of the values across the GTEx tissues}
\usage{
ggtex_boxplot_tissues(values = F, metadata, gene.rpkm = F,
  transcript.rpkm = F, gene_ids = F, transcript_ids = F,
  outlier.size = 1)
}
\arguments{
\item{values}{data.frame with the values. rownames need to be the feature identifier (gene_id, exon_id,...) and colnames the sample ids.}

\item{metadata}{object returned by load_metadata()}

\item{gene.rpkm}{gene.rpkm object; load('/users/rg/freverter/GTExAnalysis/gtex_gc19_2015_01_12/rna_seq/gene_rpkm.RData')}

\item{transcript.rpkm}{transcript.rpkm object; load('/users/rg/freverter/GTExAnalysis/gtex_gc19_2015_01_12/rna_seq/gene_rpkm.RData')}

\item{gene_ids}{list of genes}

\item{transcript_ids}{list of genes}

\item{outlier.size}{size of outlier points in the boxplot}
}
\value{
ggplot object
}
\description{
Generates a boxplot with the distribution of the values across the GTEx tissues
}
\seealso{
\code{\link{load_metadata}}
}

