% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_metadata}
\alias{load_metadata}
\title{Loads GTEx metadata}
\usage{
load_metadata(SubjectPhenotypes, SubjectSampleMapping, SampleAttributes)
}
\arguments{
\item{SubjectPhenotypes}{path to SubjectPhenotypes file.}

\item{SubjectSampleMapping}{path to SubjectSampleMapping file.}

\item{SampleAttributes}{path to SampleAttributes file.}
}
\value{
data.frame with merged metadata
}
\description{
Loads GTEx metadata
}

