% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggtex_boxplot_tissues}
\alias{ggtex_boxplot_tissues}
\title{Generates a boxplot with the distribution of the values across the GTEx tissues}
\usage{
ggtex_boxplot_tissues(values, metadata, log10_scale = T, outlier.size = 1)
}
\arguments{
\item{values}{data.frame with the values. rownames need to be the feature identifier (gene_id, exon_id,...) and colnames the sample ids.}

\item{metadata}{object obtained by load_metadata()}

\item{log10_scale}{if TRUE (default), values are transfomed by log10}

\item{outlier.size}{size of outlier points in the boxplot}
}
\value{
ggplot object
}
\description{
Generates a boxplot with the distribution of the values across the GTEx tissues
}
\seealso{
\code{\link{load_metadata}}
}

